+++ ergo/ChangeLog.md
@@ -10,6 +10,8 @@
 * Improve `ergo` use in scripts to not always apply functions when they are
   loaded. Instead, the final value at the top level will, if a function, be
   applied (with no arguments) to refine the final value.
+* Support calling `ergo` without arguments. This evaluates to the `command`
+  workspace key.
 
 ## 1.0.0-beta.4  -- 2020-11-17
 * Add the `!` operator, which may precede any expression or argument to evaluate